feat(gov): add lockup table loading and empty state
