+++ apps/governance/src/lang/extracts/en.json
@@ -170,6 +170,9 @@
   "KjK2z3": {
     "defaultMessage": "Sorry, we can't seem to find that page"
   },
+  "LpWMOQ": {
+    "defaultMessage": "No transaction"
+  },
   "MSTVqC": {
     "defaultMessage": "Reload Page"
   },
@@ -184,6 +187,9 @@
   },
   "NXI/XL": {
     "defaultMessage": "Auto"
+  },
+  "O1F5Cx": {
+    "defaultMessage": "Connect your wallet to see your history"
   },
   "OmyPDW": {
     "defaultMessage": "The length of time you will lock up your OGV in order to receive yield and voting power. There is no way to unstake before your withdrawal date."
